Also fixed ww3 output name
