+++ dnora/read/waveseries/__init__.py
@@ -1,1 +1,2 @@
-from .waveseries_readers import Spectra1DToWaveSeries
+from .waveseries_readers import Spectra1DToWaveSeries, WW3Unstruct
+from . import metno
