+++ tests/test_internal_spectra_readers.py
@@ -16,6 +16,7 @@ def timevec():
     return pd.date_range("2023-04-01 00:00:00", "2023-04-01 23:00:00", freq="1h")
 
 
+@pytest.mark.remote
 @pytest.mark.internal
 def test_wam3(grid, timevec):
     model = dn.modelrun.ModelRun(grid, year=2023, month=4, day=1)
@@ -23,6 +24,7 @@ def test_wam3(grid, timevec):
     assert np.all(model.spectra().time() == timevec)
 
 
+@pytest.mark.remote
 @pytest.mark.internal
 def test_wam800(grid, timevec):
     model = dn.modelrun.ModelRun(grid, year=2023, month=4, day=1)
@@ -30,6 +32,7 @@ def test_wam800(grid, timevec):
     assert np.all(model.spectra().time() == timevec)
 
 
+@pytest.mark.remote
 @pytest.mark.internal
 def test_nora3(grid, timevec):
     model = dn.modelrun.ModelRun(grid, year=2023, month=4, day=1)
@@ -37,6 +40,7 @@ def test_nora3(grid, timevec):
     assert np.all(model.spectra().time() == timevec)
 
 
+@pytest.mark.remote
 @pytest.mark.internal
 def test_ww3_4km(grid, timevec):
     model = dn.modelrun.ModelRun(grid, year=2023, month=4, day=1)
@@ -44,6 +48,7 @@ def test_ww3_4km(grid, timevec):
     assert np.all(model.spectra().time() == timevec)
 
 
+@pytest.mark.remote
 @pytest.mark.internal
 def test_norac(timevec):
     grid = dn.grid.Grid(lon=9.834990, lat=63.571444)
