+++ tests/test_remote_current_readers.py
@@ -17,6 +17,6 @@ def timevec():
 @pytest.mark.remote
 def test_norkyst800(grid, timevec):
     model = dn.modelrun.ModelRun(grid, year=2022, month=4, day=1)
-    model.import_current(dn.current.read_metno.NorKyst800(), program="pyfimex")
+    model.import_current(dn.read.current.metno.NorKyst800(), program="pyfimex")
 
     assert np.all(model.current().time() == timevec)
