+++ tests/test_remote_ice_readers.py
@@ -25,7 +25,7 @@ def test_nora3(grid, timevec):
     model = dn.modelrun.ModelRun(
         grid, start_time="2022-01-01 00:00", end_time="2022-01-02 23:00"
     )
-    model.import_ice(dn.ice.read_metno.NORA3())
+    model.import_ice(dn.read.ice.metno.NORA3())
     assert np.all(model.ice().time() == timevec)
 
 
@@ -34,16 +34,16 @@ def test_barents25(grid, timevec2024):
     model = dn.modelrun.ModelRun(
         grid, start_time="2024-01-05 00:00", end_time="2024-01-05 23:00"
     )
-    model.import_ice(dn.ice.read_metno.Barents25())
+    model.import_ice(dn.read.ice.metno.Barents25())
     assert np.all(model.ice().time() == timevec2024)
 
 
 @pytest.mark.remote
-def test_barents25_patc(grid):
+def test_barents25_patch(grid):
     """There are missing files (03 is missing the 18Z-folder) so we can use this to test patching"""
     model = dn.modelrun.ModelRun(
         grid, start_time="2024-01-03 00:00", end_time="2024-01-03 23:00"
     )
-    model.import_ice(dn.ice.read_metno.Barents25())
+    model.import_ice(dn.read.ice.metno.Barents25())
     timevec = pd.date_range("2024-01-03 00:00:00", "2024-01-03 23:00:00", freq="1h")
     assert np.all(model.ice().time() == timevec)
